Fix Sprockets Error on Test Suite
When attempting to run the test suite I received the following error:

    Failure/Error: RailsApp::Application.initialize!

    Sprockets::Railtie::ManifestNeededError:
      Expected to find a manifest file in `app/assets/config/manifest.js`
      But did not, please create this file and use it to link any assets that need
      to be rendered by your app:

      Example:
        //= link_tree ../images
        //= link_directory ../javascripts .js
        //= link_directory ../stylesheets .css
      and restart your server
    # ./spec/rails_app/config/environment.rb:5:in `<top (required)>'
    # ./spec/spec_helper.rb:9:in `<top (required)>'
    # ./spec/audited/sweeper_spec.rb:1:in `<top (required)>'

This is reflected on CI with the new versions of Rails failing on the
same error:

    https://travis-ci.org/github/collectiveidea/audited/jobs/686147718

Rather than create a dummy file it seemed best to just remove the
Sprockets dependency. At first I was going to include all the other
Rails libraries listed at:

    https://github.com/rails/rails/blob/master/railties/lib/rails/all.rb

I decided to cut it down further because some of these other support
libraries that we are not using may in the future cause similar issues.
Since they are not relevant to the `audited` gem it seemed good to head
off that issue.

I confirmed only the ActiveRecord support needs to be included as
long as I commented out any config related to the other components
(ActionMailer config in `test.rb`).

With ActionController now gone it seems we don't need the dummy
ApplicationController file anymore.

Finally while cleaning things up it seems we don't need development
and production environments since the test suite always runs in the
test environment.

This may all be cutting too deep. Maybe I can't trim as much with
and older version of Rails that is still supported by the `audited` gem.
If CI informs me of that I can roll back some of these cuts.