Deadlocks from concurrent audit combinations
When combining audits the oldest remaining audit is updated and the audits over max_audit count are deleted.
This can result in a deadlock if if the same record is updated at the same time. However by virtue of being at the same time
the same end is being selected and the same verion number is used for the delete. This means the same update and same delete
are being performed for both audit combinations so this error can just be ignored.