Run GitHub actions on all branches
To all for an improved contributing experience, run GitHub actions on
all branches.

When preparing a PR, I like to verify that all CI will be green on my
personal fork. However, the existing GitHub actions configuration made
it so this would not happen.

Now, contributors can run actions on their fork before opening a PR.