Relax gemspec to allow Rails 8.0; add Rails main branch to CI
Rails 8 does not have an alpha or beta rubygems release, but it sounds
like a release is imminent (to coincide with Rails World in late
September).

To prepare for Rails 8, this PR relaxes the Rails requirement in the
audited gemspec to allow "< 8.1" instead of "< 8.0". This will allow
audited to be used with Rails 8.0 when it is released.

To test that audited will work with Rails 8, I added a `rails-main`
Gemfile to `Appraisals` that targets the Rails main branch, and included
this in the CI matrix.